Fixed problem where pan/zoom didn't work after first rendering
For some reason the pan/zoom didn't work after the first time a graph
was rendered. At least if it was done at the same time the component was
created. This was fixed by initialising the pan/zooming after rendering
event has been processed completely.

Fixed formatting to to conform with the vaadin coding conventions