Fixed enclosing of id:s with double quotes
Now double quotes enclosing id-string are removed if they exist when the
node is created. That is, the id stored in a graph, node, or edge is not
surrounded by double quotes. Double quotes are added when the graph is
rendered. The reason for this design is that the graph itself would
store the "normal" string and not the string modified with double
quotes.