Support for subgraphs
Added support for subgraphs. A node can either be a regular node or a
node that itself contains a graph (i.e. a subgraph). Note that having
edges to and from a subgraph node will probably not work as expected. It
is better to add edges directly to nodes inside the subgraph.

An example illustrating use of subgraphs is also provided.